Change directions size to i64

+++ src/position/moves.rs
@@ -54,7 +54,7 @@ impl Move {
 fn generate_regular_moves_from_square(
     position: &Position,
     square: Square,
-    directions: &[i8],
+    directions: &[i64],
 ) -> Vec<Move> {
     let piece = position.at(square).unwrap();
     let color = piece.color();
@@ -64,11 +64,11 @@ fn generate_regular_moves_from_square(
 
     for offset in directions {
         let mut current_offset = *offset;
-        let mut last_column = square.col() as i8;
+        let mut last_column = square.col() as i64;
 
         loop {
-            let to_index = (square.index as i8 + current_offset) as usize;
-            let current_col = (to_index % ROW_SIZE) as i8;
+            let to_index = (square.index as i64 + current_offset) as usize;
+            let current_col = (to_index % ROW_SIZE) as i64;
             let out_of_bounds = to_index >= BOARD_SIZE || (current_col - last_column).abs() > 2;
             if out_of_bounds {
                 break;
@@ -101,7 +101,7 @@ fn pseudo_legal_moves_from_square(position: &Position, square: Square) -> Vec<Mo
 
             // Do not advance if there is a piece in front; do not capture if there is no piece
             moves.retain(|move_| {
-                let index_diff = (move_.to.index as i8 - move_.from.index as i8).abs();
+                let index_diff = (move_.to.index as i64 - move_.from.index as i64).abs();
                 if index_diff == UP {
                     !move_.flags.contains(MoveFlags::CAPTURE)
                 } else if index_diff == UP + UP {
@@ -109,7 +109,7 @@ fn pseudo_legal_moves_from_square(position: &Position, square: Square) -> Vec<Mo
                     let can_advance_two = ((row == 1 && color == Color::White)
                         || (row == 6 && color == Color::Black))
                         && !move_.flags.contains(MoveFlags::CAPTURE);
-                    let jumped_piece = position.board[(move_.from.index as i8
+                    let jumped_piece = position.board[(move_.from.index as i64
                         + (if color == Color::White { UP } else { DOWN }))
                         as usize]
                         .is_some();
@@ -276,8 +276,8 @@ pub fn make_move(position: &Position, move_: Move) -> Position {
     // En passant
     if move_.flags.contains(MoveFlags::ENPASSANT) {
         let capture_square = match position.to_move {
-            Color::White => move_.to.index as i8 + DOWN,
-            Color::Black => move_.to.index as i8 + UP,
+            Color::White => move_.to.index as i64 + DOWN,
+            Color::Black => move_.to.index as i64 + UP,
         };
         new_board[capture_square as usize] = None;
     }
@@ -346,7 +346,7 @@ pub fn make_move(position: &Position, move_: Move) -> Position {
         castling_rights: new_castling_rights,
         en_passant_square: match position.at(move_.from).unwrap() {
             Piece::WP | Piece::BP => {
-                if (move_.to.index as i8 - move_.from.index as i8).abs() == 2 * UP {
+                if (move_.to.index as i64 - move_.from.index as i64).abs() == 2 * UP {
                     Some(Square { index: (move_.to.index + move_.from.index) / 2 })
                 } else {
                     None
